move LootMenu
