+++ src/XSEPlugin.cpp
@@ -3,8 +3,8 @@
 #include "Animation/Animation.h"
 #include "Events/Events.h"
 #include "Hooks.h"
+#include "LootMenu.h"
 #include "LootMenuManager.h"
-#include "Scaleform/LootMenu.h"
 #include "Integrations/LOTD.h"
 #include "Integrations/Completionist.h"
 #include "Integrations/APIServer.h"
